requirements.txt
numpy==1.16.4
h5py==2.8.0
opencv-python==4.1.0.25
Pillow==5.0.0
scipy==1.1.0
tensorflow-gpu>=1.14.1
gpustat
scikit-image


